Separate worker process and add supervisor ability
By default, all the work is done in a child process. The main process
waits for the child process and cleans up the HW state in case the child
process does not exit cleanly. The main purpose is to make sure that
heating/cooling are turned off if the worker process crashes.

Additional changes:
  * Improved command line parsing.
  * Improved HW initialization error checking.