Re-enable the secondary sensor
The secondary sensor is read, but the readings are still not used as
input to the HVAC control algorithm. The hard-coded calibration
constants for the secondary sensor have been removed. They seem to be
unnecessary - at least the temperature one.

The delay between reading the main sensor and the secondary sensor has
been changed from 10ms to 200ms. It was determined experimentally that a
delay of 197ms is too little and the Modbus transaction times out, and a
delay of 197.5ms works. The reason is currently unknown.