Simplify network code by using blocking mode for sockets
Portability is not a goal, since other Linux specific features are
already used (e.g. timerfd) in other parts of the project. Use blocking
mode for sockets and the Linux specific SO_SNDTIMEO and SO_RCVTIMEO
options to implement timeout options. The code is simplified compared
to poll(), which otherwise is not needed, since there is no use case for
waiting on multiple file descriptors simultaneously.