Fix common macros that include fallback statements
Using a dummy "do/while" loop is the universally accepted method for
creating macros that contain "if" statements and are safe to use inside
other "if/else" statements. However, they can introduce very subtle bugs
if the fallback instruction (which is a macro parameter) happens to be a
"break" instruction, because it will break out of the dummy "do/while"
loop inside the macro instead of the outer loop surrounding the macro
call site (which is what the user would expect).

Use compound statement expressions, which are a GNU/gcc extension, since
they solve both issues and are already used in other spots. Portability
is not a concern here.