Add a startup delay for the ERV
After the program starts, it takes roughly two seconds to connect to
memcache and start retrieving the outdoor temperature data. If the ERV
control algorithm starts running immediately, before the memcache data
is available, it will notice that the data is stale and set the ERV mode
to i30mh. Then, once the memcache data becomes available, if the outdoor
temperature is high enough to run the ERV continuously, changing the ERV
mode will be delayed by the ERV holdoff timer.

Avoid the initial ERV holdoff timer delay by adding a startup delay to
the ERV algorithm, to avoid looking at the outdoor temperature before
the memcache data is available.