Adjust ERV control for furnace fan mode
When the furnace is in "fan" mode, there is no forced heating or
cooling, so the ERV must run at lower speeds to preserve the indoors
temperature.

Create an additional temperature offset map for furnace fan mode, and
use the average of the heat and cool setpoints as the reference.
Previously the indoors temperature was used as the reference.