Add support for sensor calibration data
Sensor calibration data is defined in the configuration file and is
sparse. The full structure is added to config.json.sample, but any parts
of it (or the whole structure) can be missing. The parser is relaxed
thanks to graceful NULL pointer handling throughout the cJSON library.
The calibration values that are not explicitly defined default to 0.