Send ERV actual mode and humidifier state to Graphite
When the ERV mode is configured to "auto", the actual ERV mode is
determined by the control algorithm. Looking at historical data in
Graphite, it may be useful to know what mode the ERV really ran in.
This data can be (partially) inferred from the ERV control pins but
it's not very straightforward. Log the actual ERV mode, in addition
to the control pins state.

Similarly, the time when the humidifier was on can be inferred from
the valve control pin, which is active intermittently in 30 seconds
cycles. However, that is much too granular. Log the actual ON/OFF
state, in addition to the control pins state.