Avoid humidifier state glitch on sensor read failure
Reading the sensors may fail e.g. due to a modbus error. In that case,
sensor_read() returns with an error but humid_update() is still called.
That's correct because valve pulse control and auxiliary fan control
still need to occur. But the humidifier state must not be changed based
on invalid sensor data.